// src/@types/account-type.enum.ts
export enum AccountType{
    SALES = 'sales',
    OVERHEADS =  "overheads",
    CURRENT_ACCOUNTS_RECEIVABLE = "current_accounts_receivable",
    BANK = "bank",
    FIXED = "fixed",
    CURRENT_ACCOUNTS_PAYABLE = "current_accounts_payable",
    TAX = "tax",
    CURRENT= "current",
    PAYROLL =  "payroll",
}